Convert asyncify tests to mocha
