+++ test/test-async.js
@@ -16,11 +16,6 @@ if (!Function.prototype.bind) {
     };
 }
 
-function isBrowser() {
-    return (typeof process === "undefined") ||
-        (process + "" !== "[object process]"); // browserify
-}
-
 exports['sortBy'] = function(test){
     test.expect(2);
 
@@ -575,108 +570,4 @@ exports['constant'] = function (test) {
         test.ok(c === 3);
         test.done();
     });
-};
-
-exports['asyncify'] = {
-    'asyncify': function (test) {
-        var parse = async.asyncify(JSON.parse);
-        parse("{\"a\":1}", function (err, result) {
-            test.ok(!err);
-            test.ok(result.a === 1);
-            test.done();
-        });
-    },
-
-    'asyncify null': function (test) {
-        var parse = async.asyncify(function() {
-            return null;
-        });
-        parse("{\"a\":1}", function (err, result) {
-            test.ok(!err);
-            test.ok(result === null);
-            test.done();
-        });
-    },
-
-    'variable numbers of arguments': function (test) {
-        async.asyncify(function (x, y, z) {
-            test.ok(arguments.length === 3);
-            test.ok(x === 1);
-            test.ok(y === 2);
-            test.ok(z === 3);
-        })(1, 2, 3, function () {});
-        test.done();
-    },
-
-    'catch errors': function (test) {
-        async.asyncify(function () {
-            throw new Error("foo");
-        })(function (err) {
-            test.ok(err);
-            test.ok(err.message === "foo");
-            test.done();
-        });
-    },
-
-    'dont catch errors in the callback': function (test) {
-        try {
-            async.asyncify(function () {})(function (err) {
-                if (err) {
-                    return test.done(new Error("should not get an error here"));
-                }
-                throw new Error("callback error");
-            });
-        } catch (e) {
-            test.ok(e.message === "callback error");
-            test.done();
-        }
-    },
-
-    'promisified': [
-        'native-promise-only',
-        'bluebird',
-        'es6-promise',
-        'rsvp'
-    ].reduce(function(promises, name) {
-        if (isBrowser()) {
-            // node only test
-            return;
-        }
-        var Promise = require(name);
-        if (typeof Promise.Promise === 'function') {
-            Promise = Promise.Promise;
-        }
-        promises[name] = {
-            'resolve': function(test) {
-                var promisified = function(argument) {
-                    return new Promise(function (resolve) {
-                        setTimeout(function () {
-                            resolve(argument + " resolved");
-                        }, 15);
-                    });
-                };
-                async.asyncify(promisified)("argument", function (err, value) {
-                    if (err) {
-                        return test.done(new Error("should not get an error here"));
-                    }
-                    test.ok(value === "argument resolved");
-                    test.done();
-                });
-            },
-
-            'reject': function(test) {
-                var promisified = function(argument) {
-                    return new Promise(function (resolve, reject) {
-                        reject(argument + " rejected");
-                    });
-                };
-                async.asyncify(promisified)("argument", function (err) {
-                    test.ok(err);
-                    test.ok(err.message === "argument rejected");
-                    test.done();
-                });
-            }
-        };
-        return promises;
-    }, {})
 };
